feat(disk_space_analyzer): optimizing the disk analysis research

+++ antarest/study/model.py
@@ -42,6 +42,7 @@ from sqlalchemy import (
     String,
 )
 from sqlalchemy.orm import Mapped, mapped_column, relationship, validates
+from sqlalchemy.sql.sqltypes import BigInteger
 from typing_extensions import override
 
 from antarest.core.model import PublicMode
@@ -457,7 +458,7 @@ class StudyDiskSpaceAnalysis(Base):
         nullable=False,
         primary_key=True,
     )
-    disk_space_bytes: Mapped[int] = mapped_column(Integer, nullable=False)
+    disk_space_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False)
     last_analysis_date: Mapped[datetime] = mapped_column(DateTime, nullable=False)
     study: Mapped["Study"] = relationship("Study", uselist=False)
